Add app manifest + catalog synths (#28 Phase 1)
